requirements.txt
# ==================================
# Core Deep Learning
# ==================================
# Note: PyTorch should be installed via Conda for GPU support
# torch
# torchvision
# torchaudio

# ==================================
# Training & Logging
# ==================================
tensorboard>=2.12.0
tqdm>=4.65.0

# ==================================
# Data Processing & Metrics
# ==================================
scikit-learn>=1.2.0
numpy>=1.24.0
pandas>=2.0.0

# ==================================
# Visualization
# ==================================
matplotlib>=3.7.0
seaborn>=0.12.0

# ==================================
# Configuration
# ==================================
pyyaml>=6.0

# ==================================
# Image Processing
# ==================================
opencv-python>=4.8.0
pillow>=10.0.0

# ==================================
# GUI
# ==================================
customtkinter>=5.2.0
